Floating the ints

+++ Assets/Scripts/Bars/HealthBar.cs
@@ -10,7 +10,7 @@ public class HelathBar : MonoBehaviour
 
     public GameObject playerState;
 
-    private int currentHealth, maxHealth;
+    private float currentHealth, maxHealth;
     void Awake() {
         slider = GetComponent<Slider>();
     }
@@ -23,6 +23,6 @@ public class HelathBar : MonoBehaviour
         float fillValue = (float)currentHealth / (float)maxHealth;  //used to calculate slider value to display health
         slider.value = fillValue;
 
-        healthCounter.text = currentHealth + "/" + maxHealth;
+        healthCounter.text = (int)currentHealth + "/" + (int)maxHealth;
     }
 }
